Feedback request added / updated
The DLL has been changed and CmdSwitchInfoRequest(..) was added. The examples were updated and are now able to request the correct feedback data from the command station. In the examples the default IP addresses have been changed to the default Lenz and Roco addresses. The switch direction changed (1 is now straigth). Some minor updates regarding documentation.