Instructions to run the XpressNet code WITHIN your VB6 code improved
The previous description of how to integrate the XpressNet code within your own application incorrect, in the sense that is proposed to create a VB group. Although such VB groups are very useful to debug the XpressNet code, such groups can not compile the code into a single .exe. Therefore the descriptions were corrected and improved.